backend/requirements.txt
django==4.2.17
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==11.1.0
